--- recommend-service/src/main/java/com/sparta/recommendservice/domain/product_vector/application/PythonRunner.java
@@ -1,7 +1,0 @@
-package com.sparta.recommendservice.domain.product_vector.application;
-
-import java.io.IOException;
-
-public interface PythonRunner {
-	void run() throws InterruptedException, IOException;
-}
